Create new rollup config
TODO: migrate deployment scripts, functional tests, and custom build
