+++ test/unit/specs/components/Personalization/validateApplyPropositionsOptions.spec.js
@@ -11,7 +11,7 @@ governing permissions and limitations under the License.
 */
 
 import { vi, beforeEach, describe, it, expect } from "vitest";
-import { DOM_ACTION } from "@adobe/alloy/libEs6/constants/schema.js";
+import { DOM_ACTION } from "../../../../../src/constants/schema.js";
 import validateApplyPropositionsOptions, {
   EMPTY_PROPOSITIONS,
 } from "../../../../../src/components/Personalization/validateApplyPropositionsOptions.js";
