+++ tsconfig.json
@@ -5,7 +5,7 @@
     "allowJs": true,
     "declaration": true,
     "emitDeclarationOnly": true,
-    "outDir": "types",
+    "outDir": "dist/",
     "declarationMap": true,
   }
 }
